add nav links for desktop version need to add css style later
