+++ src/components/nav/index.js
@@ -1,6 +1,8 @@
-import React, {Component} from 'react';
+import React, {Component, Fragment} from 'react';
 import SideNav from './sidenav';
 import './nav.scss';
+import {Link} from 'react-router-dom';
+import {connect} from 'react-redux';
 
 class Nav extends Component {
     constructor(props){
@@ -10,7 +12,7 @@ class Nav extends Component {
             sideNav: {
                 body:{width: 0},
                 background:{width: 0} 
-            }
+            },
         };
     }
 
@@ -34,7 +36,109 @@ class Nav extends Component {
         }
     }
 
+    signInUser() {
+        const {trips_id} = this.props;
+
+        if(trips_id) {
+            return[
+                {
+                    to: '/mytrip',
+                    text: 'Current Trip'
+                },
+                {
+                    to: '/previoustrips',
+                    text: 'Previous Trips'
+                },
+                {
+                    to: '/account/signout',
+                    text: 'Sign Out'
+                }
+            ];
+        } else {
+            return[
+                {
+                    to: '/',
+                    text: 'Home'
+                },
+                {
+                    to: '/previoustrips',
+                    text: 'Previous Trips'
+                },
+                {
+                    to: '/account/signout',
+                    text: 'Sign Out'
+                }
+            ];
+        }
+    }
+
+    guest() {
+        const {trips_id} = this.props;
+
+        if(trips_id) {
+            return [
+                {
+                    to: '/mytrip',
+                    text: 'Current Trip'
+                },
+                {
+                    to: '/account/signin',
+                    text: 'Sign In'
+                },
+                {
+                    to: '/account/signup',
+                    text: 'Sign Up'
+                }
+            ];
+        } else {
+            return [
+                {
+                    to: '/',
+                    text: 'Home'
+                },
+
+                {
+                    to: '/account/signin',
+                    text: 'Sign In'
+                },
+                {
+                    to: '/account/signup',
+                    text: 'Sign Up'
+                }
+            ];
+        }
+    }
+
+    renderLinks() {
+        const {guest, userAuth} = this.props;
+        let navlink = null;
+
+        if(!guest && userAuth) {
+            navlink = this.signInUser();
+        } else {
+            navlink = this.guest();
+        }
+
+        let links = navlink.map(this.buildLink);
+
+        return (
+            <Fragment>
+              {links}  
+            </Fragment>
+        );
+    }
+
+    buildLink(link) {
+        return (
+            <li key={link.to}><Link to={link.to}>{link.text}</Link></li>
+        );
+    }
+
     render() {
+        const links = this.renderLinks();
+        const {sideNav, greeting} = this.state;
+
+
         return (
             <div className="nav-box">
                 <nav className="navbar">
@@ -46,11 +150,22 @@ class Nav extends Component {
                     <div className="nav-title">
                         <h1>MysterTravel</h1>
                     </div>
+                    <div className="navbar-links">
+                        {links}
+                    </div>
                 </nav>
-                <SideNav style={this.state.sideNav} toggle={this.toggleSideNav}/>
+                <SideNav style={sideNav} toggle={this.toggleSideNav} links={links} greeting={greeting}/>
             </div>
         );
     }
 }
 
-export default Nav;
+function mapStateToProps(state){
+    return{
+        trips_id: state.user.trips_id,
+        userAuth: state.user.auth,
+        guest: state.user.guest
+    }
+}
+
+export default connect(mapStateToProps)(Nav);
